requirements.txt
PySide6>=6.6
pyqtgraph>=0.13
numpy>=1.26
scipy>=1.11
soundfile>=0.12
pydub>=0.25
Pillow>=10.0
imageio-ffmpeg>=0.4.9
pyinstaller>=6.0
pytest>=7.4
